requirements.txt
streamlit
requests
python-dotenv
pandas
googlesearch-python
